Introduce adjacent_to_void/rail masks
Computing the masks, even if unused, slows the program down.  I assume
this is just randomness in code layout or something.

Using the masks speeds the program up from that baseline.

It seems like we should sink the initialization of next until after
checking the new masks, but doing so both slows the program down and
increases the number of instructions executed (so probably a real
effect).