Write separate start and length files
To query, use std::upper_bound to find one past the start of the
interval, then decrement that value (if it's the first, fail query).
Then load the corresponding byte from the length file to check if the
value is within that interval.

Using a separate length file makes the query a little simpler and also
should make both files more compressible than the merged file.