Remove LMDB; resolve outcounts instead of flushing
LMDB can't keep up with multithreaded writes.  It seems that duplicate
win-loss queries are local, so we can get most of the benefit by just
batching.  (Though it raises the question of why we were taking so many
page faults in the first place, if there is locality in requests.  Maybe
the working set is too big when we have 64 threads?)